requirements_Version2.txt
azure-ai-documentintelligence>=1.0.0
azure-core>=1.29.0
python-dotenv>=1.0.1